Changes:
- Skip deprecated entities and top-level entities (owl:thing and top
data/object property).
- Check if entities in the ontology signature have annotations in
imported ontologies. Before this, if a class from an imported ontology
was used, for example in a subclass axiom, that class is in the
signature of the evaluated ontology, but the annotations of that class
were not included as they belong to the imported ontology, thus
considering an empty class.